requirements.txt
flask
python-dotenv
requests
torch
transformers
openai-whisper
librosa
scikit-learn
pyttsx3
sounddevice
soundfile
numpy
openai-whisper
edge-tts
playsound3